feat: add function of getting clinet id and user token in /store
